pyproject.toml
[project]
name = "jepa-intuitive-physics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beartype>=0.22.2",
    "braceexpand>=0.1.7",
    "decord>=0.6.0",
    "einops>=0.8.1",
    "fabric>=3.2.2",
    "fire>=0.7.1",
    "ftfy>=6.3.1",
    "h5py>=3.14.0",
    "hydra-core>=1.3.2",
    "iopath>=0.1.10",
    "jaxtyping>=0.3.3",
    "lightning>=2.5.5",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "peft>=0.17.1",
    "psutil>=7.1.0",
    "python-box>=7.3.2",
    "pyyaml>=6.0.3",
    "rootutils>=1.0.7",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "submitit>=1.5.3",
    "tensorboard>=2.20.0",
    "timm>=1.0.20",
    "torch>=2",
    "torchvision>=0.23.0",
    "transformers>=4.57.0",
    "wandb>=0.22.2",
    "webdataset>=1.0.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.0.1",
    "pip>=25.2",
]

[tool.ruff.lint]
ignore = ["F722"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Defines the targets for the wheel build.
packages = ["evaluation_code/src", "intphyseval"]
